# apilayer/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    Location:
        properties:
            lat:
                format: double
                type: number
                x-go-name: Lat
            lon:
                format: double
                type: number
                x-go-name: Lon
        title: Location ...
        type: object
        x-go-package: github.com/mohit2530/communityCare/model
...
    MessageResponse:
        properties:
            Message:
                type: string
        title: MessageResponse ...
        type: object
        x-go-package: github.com/mohit2530/communityCare
    Notes:
        properties:
            color:
                type: string
                x-go-name: Color
            completionDate:
                format: date-time
                type: string
                x-go-name: CompletionDate
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            created_by:
                type: string
                x-go-name: CreatedBy
            creator:
                type: string
                x-go-name: Creator
            description:
                type: string
                x-go-name: Description
            location:
                $ref: '#/definitions/Location'
            noteID:
                type: string
                x-go-name: ID
            sharable_groups:
                items:
                    format: uuid
                    type: string
                type: array
                x-go-name: SharableGroups
            status:
                type: string
                x-go-name: Status
            status_description:
                type: string
                x-go-name: StatusDescription
            status_name:
                type: string
                x-go-name: StatusName
            title:
                type: string
                x-go-name: Title
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
            updated_by:
                type: string
                x-go-name: UpdatedBy
            updator:
                type: string
                x-go-name: Updator
        title: Notes ...
        type: object
        x-go-name: Note
        x-go-package: github.com/mohit2530/communityCare/model
...
    Profile:
        properties:
            activity_id:
                type: string
                x-go-name: ActivityID
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            created_by:
                type: string
                x-go-name: CreatedBy
            creator:
                type: string
                x-go-name: Creator
            custom_action:
                type: string
                x-go-name: CustomAction
            id:
                format: uuid
                type: string
                x-go-name: ID
            sharable_groups:
                items:
                    type: string
                type: array
                x-go-name: SharableGroups
            title:
                type: string
                x-go-name: Title
            type:
                type: string
                x-go-name: Type
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
            updated_by:
                type: string
                x-go-name: UpdatedBy
            updator:
                type: string
                x-go-name: Updator
        title: Profile ...
        type: object
        x-go-name: RecentActivity
        x-go-package: github.com/mohit2530/communityCare/model
...
info:
    description: |-
        # Documentation for the Asset alert api layer.

        Scehmes: http
    title: alert Api Layer
    version: 1.0.0
paths:
...
    /api/profile/{id}/notes:
        delete:
            operationId: removeNote
            parameters:
                - description: The id of the selected note
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Removes the note from the db
            tags:
                - RemoveNote
        get:
            operationId: getNotes
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: Notes
                    schema:
                        items:
                            $ref: '#/definitions/Notes'
                        type: array
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Retrieves the list of notes for the selected user
            tags:
                - GetNotes
        post:
            operationId: addNewNote
            parameters:
                - description: The id of the selected note
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The note object to add into the db
                  in: body
                  name: Note
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The note object to add into the db
                    type: object
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Add a new note to the database
            tags:
                - AddNewNote
        put:
            operationId: updateNote
            parameters:
                - description: The id of the selected note
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The note object to update into the db
                  in: body
                  name: Note
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The note object to update into the db
                    type: object
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Updates an existing note in the database
            tags:
                - UpdateNote
...
    /api/v1/profile:
        put:
            operationId: updateProfile
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The full profile object of the user
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/Profile'
                    description: The full profile object of the user
                    type: object
            responses:
                "200":
                    description: Profile
                    schema:
                        $ref: '#/definitions/Profile'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Updates the current user profile for the selected user. Does not meddle with authentication
            tags:
                - UpdateProfile
    /api/v1/profile/{id}:
        get:
            operationId: getUsername
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Retrieves the user name from the profiles table. Does not meddle with authentication
            tags:
                - GetUsername
...
    /api/v1/profile/{id}/updateAvatar:
        post:
            operationId: updateProfileAvatar
            parameters:
                - description: The userID of the selected user
                  in: path
                  required: true
                  type: string
                - description: The full file details of the avatar
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FileHeader'
                    description: The full file details of the avatar
                    type: object
            responses:
                "200":
                    description: Profile
                    schema:
                        $ref: '#/definitions/Profile'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Updates the current user avatar for the selected user. Does not meddle with authentication
            tags:
                - UpdateProfileAvatar
    /api/v1/profile/list:
        get:
            operationId: getAllUserProfiles
            responses:
                "200":
                    description: Profile
                    schema:
                        items:
                            $ref: '#/definitions/Profile'
                        type: array
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Retrieves all the user profiles listed in the database
            tags:
                - GetAllUserProfiles
...
produces:
    - application/json
swagger: "2.0"